distroboot: Always check for custom boot scripts first
If extlinux.conf is used, then it's not possible to customize boot
environment, because scripts are not loaded.
Usually it's possible to make some changes manually using command line
and save boot environment. But if exlinux.conf is loaded
from ext4 partition (for example on PinePhone), then environment are not saved/loaded at boot
time from boot partition and it's not possible to persistently change
boot environment without recompiling u-boot.

Signed-off-by: Andrey Skvortsov <andrej.skvortzov@gmail.com>